Add minute callback implementation to static temp window.
Also refactor temperature from uint32_t to float.

+++ inc/static_temp_window.hpp
@@ -28,10 +28,10 @@ public:
 	virtual void minCallback() override;
 	void hide() override;
 	void show() override;
-	void setTemp(uint32_t temp);
+	void setTemp(double temp);
 	void measure();
 private:
-	uint32_t temp = 0;
+	double temp = 0;
 	bool registeredCallback = false;
 };
